docs/journal-backend-bootstrap-pgvector-fts-reindex/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "journal-backend"
version = "0.1.0"
description = "FastAPI backend scaffold (Sept 2025)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "openai>=1.30.0",
  "pgvector>=0.2.5",
  "fastapi[standard]>=0.115.0",        # keep semver flexible; use latest 0.11x
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "sqlalchemy[postgresql,asyncio]>=2.0.25",
  "psycopg[binary,pool]>=3.2",
  "asyncpg>=0.29.0",
  "sqlmodel>=0.0.21",
  "alembic>=1.13.0",
  "redis>=5.0.0",
  "pyjwt>=2.8.0",
  "authlib>=1.3.0",
  "httpx>=0.27.0",
  "opentelemetry-api>=1.25.0",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp>=1.25.0",
  "opentelemetry-instrumentation-fastapi>=0.50b0",
  "nats-py>=2.7.0",
  "strawberry-graphql>=0.200.0",
  "python-multipart>=0.0.9",
  "ruff>=0.5.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-select = ["I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
